+++ src/Ingredients/Ingredients.API/Entities/Ingredients.cs
@@ -5,10 +5,10 @@ namespace src.Ingredients.Ingredients.API
         public Guid Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public decimal Calories { get; set; }
-        public decimal Proteins { get; set; }
-        public decimal Carbs { get; set; }
-        public decimal Fats { get; set; }
+        public double Calories { get; set; }
+        public double Proteins { get; set; }
+        public double Carbs { get; set; }
+        public double Fats { get; set; }
     }
 
 }
